+++ config/services.php
@@ -35,4 +35,11 @@ return [
         ],
     ],
 
+    'mercadopago' => [
+        'access_token' => env('MERCADOPAGO_ACCESS_TOKEN'),
+        'webhook_secret' => env('MERCADOPAGO_WEBHOOK_SECRET'),
+        'webhook_url' => env('MERCADOPAGO_WEBHOOK_URL'),
+        'return_url' => env('MERCADOPAGO_RETURN_URL', env('APP_URL')),
+    ],
+
 ];